parsers/playwright_22bet/requirements.txt
playwright==1.40.0
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python==8.0.33
